refactoring lecture2_task3 made directory for the task
