+++ homework/java/natal-lia-b/hw1/projectHW1/src/main/java/by/jrr/start/lecture2_task3/Addition.java
@@ -1,4 +1,4 @@
-package by.jrr.start;
+package by.jrr.start.lecture2_task3;
 
 class Addition {
     Addition(int a, int b) {
