+++ homework/java/natal-lia-b/hw1/projectHW1/src/main/java/by/jrr/start/lecture2_task3/MathOperationsDemo.java
@@ -1,4 +1,4 @@
-package by.jrr.start;
+package by.jrr.start.lecture2_task3;
 /*
 Описание:
 Разработать программу, которая демонстрирует работу с простейшими математическими операциями.
