+++ homework/java/natal-lia-b/hw1/projectHW1/src/main/java/by/jrr/start/lecture2_task3/Mod.java
@@ -1,4 +1,4 @@
-package by.jrr.start;
+package by.jrr.start.lecture2_task3;
 
 class Mod {
     Mod(int a, int b) {
